Create a new class hierarchy for assembly parsers.
This commit pulls all of the Mach-O specific bits out of lib/MC and puts
them into a MachOAsmParser class. This allows MachOAsmParser to take
care of implementing all the Mach-O assembler directives and the only
directives that are parsed in AsmParser are target-independent.
TargetAsmParser is deleted and all targets should interit from the
object file that the asm parser understands (currently just
MachOAsmParser).

Note that there's a missing commit to clang (because this repo doesn't
mirror the clang/ source).